Перенес логику отрисовки и сохранения в новые классы
