--- cs/TagsCloudVisualization/VisualizationCircularCloudLayout.cs
@@ -1,47 +1,0 @@
-﻿using System.Drawing;
-using System.Drawing.Imaging;
-
-namespace TagsCloudVisualization;
-#pragma warning disable CA1416
-
-public class VisualizationCircularCloudLayout
-{
-    public static Bitmap DrawLayout(List<Rectangle> rectangles, int paddingFromBorders)
-    {
-        var minX = rectangles.Min(rectangle => rectangle.Left);
-        var minY = rectangles.Min(rectangle => rectangle.Top);
-        var maxX = rectangles.Max(rectangle => rectangle.Right);
-        var maxY = rectangles.Max(rectangle => rectangle.Bottom);
-        var width = maxX - minX + paddingFromBorders;
-        var height = maxY - minY + paddingFromBorders;
-
-        var random = new Random();
-        
-        var bitmap = new Bitmap(width, height);
-        using var graphics = Graphics.FromImage(bitmap);
-        
-        graphics.Clear(Color.White);
-        
-        foreach (var rectangle in rectangles)
-        {
-            var shiftedRectangle = rectangle with
-            {
-                X = rectangle.X - minX + paddingFromBorders, 
-                Y = rectangle.Y - minY + paddingFromBorders
-            };
-            
-            var randomColor = Color.FromArgb(random.Next(256), random.Next(256), random.Next(256));
-            
-            var brush = new SolidBrush(randomColor);
-            graphics.FillRectangle(brush, shiftedRectangle);
-            graphics.DrawRectangle(Pens.Black, shiftedRectangle);
-        }
-
-        return bitmap;
-    }
-
-    public static void SaveLayout(Bitmap bitmap, string filePath, string fileName, ImageFormat imageFormat)
-    {
-        bitmap.Save(Path.Combine(filePath, fileName), imageFormat);
-    }
-}
